fix(ipc): 补齐收口最终步骤（main.rs invoke_handler 切换 + lsp version u32）
上一提交 28c1f04 漏推这两个文件，导致：collect_commands! 已含 LSP 但 lsp_did_change 的 version 仍为 i64，debug 下 export() 会触发 BigInt-forbidden panic；且 main.rs 未切换。

- lsp/commands: lsp_did_change 参数 version i64 -> u32（导出安全；文档版本号单调正增，零行为改动）
- main.rs: .invoke_handler(specta_bindings.invoke_handler()) 置于 .setup() 之前；删除手写 generate_handler!；清理仅供 generate_handler! 使用的命令导入，仅保留 shutdown_all_terminal_sessions / TerminalSessionState。

至此 collect_commands!(104) = 原 generate_handler!(104)，运行时命令注册单一来源，Rust 侧双轨彻底收口。

+++ src-tauri/src/commands/lsp/commands.rs
@@ -289,7 +289,7 @@ pub async fn lsp_did_change(
     manager: tauri::State<'_, LspManager>,
     file_path: String,
     content: String,
-    version: i64,
+    version: u32,
 ) -> Result<(), String> {
     let (stdin, uri, _) = require_running_with_uri(&manager, &file_path, false).await?;
     let params = serde_json::json!({
